updates to js and creation of module for hide/show
